updated the real-human record version

+++ download_resources.sh
@@ -34,6 +34,7 @@ echo "Downloading and extracting HRSP2mix..."
 # Default values
 sample_rate="all"  # First arg: 8k, 16k, or all
 version="all"      # Second arg: raw, clean, or all
+type="all"         # Third arg: tts, real, or all
 
 # Check command line arguments
 if [ $# -ge 1 ]; then
@@ -42,34 +43,74 @@ fi
 if [ $# -ge 2 ]; then
     version="$2"
 fi
+if [ $# -ge 3 ]; then
+    type="$3"
+fi
 
 # Download based on sample rate and version
 if [[ "$sample_rate" == "8k" || "$sample_rate" == "all" ]]; then
-    mkdir -p HRSP2mix_8k/
-    if [[ "$version" == "raw" || "$version" == "all" ]]; then
-        # wget -O temp_8k_raw.tar.gz https://archive.org/download/hrsp-2mix-8000-raw/HRSP2mix_8000_raw.zip
-        wget -O https://archive.org/download/hrsp-2mix-8000-human-raw/HRSP2mix_8000_human_raw.zip
-        tar -xzf temp_8k_raw.tar.gz -C /HRSP2mix_8k/
-        rm temp_8k_raw.tar.gz
+    
+    # tts
+    if [[ "$type" == "tts" || "$type" == "all" ]]; then
+        mkdir -p HRSP2mix_8000_human/
+        if [[ "$version" == "raw" || "$version" == "all" ]]; then
+            # wget -O temp_8k_raw.tar.gz https://archive.org/download/hrsp-2mix-8000-raw/HRSP2mix_8000_raw.zip
+            wget -O https://archive.org/download/hrsp-2mix-8000-human-raw/HRSP2mix_8000_human_raw.zip
+            tar -xzf temp_8k_raw.tar.gz -C /HRSP2mix_8000_human/
+            rm temp_8k_raw.tar.gz
+        fi
+        if [[ "$version" == "clean" || "$version" == "all" ]]; then
+            wget -O temp_8k_clean.tar.gz https://archive.org/download/hrsp-2mix-8000-human-clean-0512/HRSP2mix_8000_human_clean.zip
+            tar -xzf temp_8k_clean.tar.gz -C HRSP2mix_8000_human/
+            rm temp_8k_clean.tar.gz
+        fi
     fi
-    if [[ "$version" == "clean" || "$version" == "all" ]]; then
-        wget -O temp_8k_clean.tar.gz https://archive.org/download/hrsp-2mix-8000-human-clean-0512/HRSP2mix_8000_human_clean.zip
-        tar -xzf temp_8k_clean.tar.gz -C HRSP2mix_8k/
-        rm temp_8k_clean.tar.gz
+
+    # real
+    if [[ "$type" == "real" || "$type" == "all" ]]; then
+        mkdir -p HRSP2mix_8000_real_human/
+        if [[ "$version" == "raw" || "$version" == "all" ]]; then
+            wget -O temp_8k_raw.tar.gz https://archive.org/download/hrsp-2mix-8000-real-human-raw/HRSP2mix_8000_real_human_raw.zip
+            tar -xzf temp_8k_raw.tar.gz -C /HRSP2mix_8000_real_human/
+            rm temp_8k_raw.tar.gz
+        fi
+        if [[ "$version" == "clean" || "$version" == "all" ]]; then
+            wget -O temp_8k_clean.tar.gz https://archive.org/download/hrsp-2mix-8000-real-human-clean/HRSP2mix_8000_real_human_clean.zip
+            tar -xzf temp_8k_clean.tar.gz -C HRSP2mix_8000_real_human/
+            rm temp_8k_clean.tar.gz
+        fi
     fi
-fi
 
 if [[ "$sample_rate" == "16k" || "$sample_rate" == "all" ]]; then
-    mkdir -p HRSP2mix_16k/
-    if [[ "$version" == "raw" || "$version" == "all" ]]; then
-        wget -O temp_16k_raw.tar.gz https://archive.org/download/hrsp-2mix-16000-human-raw-0512/HRSP2mix_16000_human_raw.zip
-        tar -xzf temp_16k_raw.tar.gz -C HRSP2mix_16k/
-        rm temp_16k_raw.tar.gz
+    
+    # tts
+    if [[ "$type" == "tts" || "$type" == "all" ]]; then
+        mkdir -p HRSP2mix_16000_human/
+        if [[ "$version" == "raw" || "$version" == "all" ]]; then
+            wget -O temp_16k_raw.tar.gz https://archive.org/download/hrsp-2mix-16000-human-raw-0512/HRSP2mix_16000_human_raw.zip
+            tar -xzf temp_16k_raw.tar.gz -C HRSP2mix_16000_human/
+            rm temp_16k_raw.tar.gz
+        fi
+        if [[ "$version" == "clean" || "$version" == "all" ]]; then
+            wget -O temp_16k_clean.tar.gz https://archive.org/download/hrsp-2mix-16000-human-clean-0512/HRSP2mix_16000_human_clean.zip
+            tar -xzf temp_16k_clean.tar.gz -C HRSP2mix_16000_human/
+            rm temp_16k_clean.tar.gz
+        fi
     fi
-    if [[ "$version" == "clean" || "$version" == "all" ]]; then
-        wget -O temp_16k_clean.tar.gz https://archive.org/download/hrsp-2mix-16000-human-clean-0512/HRSP2mix_16000_human_clean.zip
-        tar -xzf temp_16k_clean.tar.gz -C HRSP2mix_16k/
-        rm temp_16k_clean.tar.gz
+
+    # real
+    if [[ "$type" == "real" || "$type" == "all" ]]; then
+        mkdir -p HRSP2mix_16000_real_human/
+        if [[ "$version" == "raw" || "$version" == "all" ]]; then
+            wget -O temp_16k_raw.tar.gz https://archive.org/download/hrsp-2mix-16000-real-human-raw/HRSP2mix_16000_real_human_raw.zip
+            tar -xzf temp_16k_raw.tar.gz -C HRSP2mix_16000_real_human/
+            rm temp_16k_raw.tar.gz
+        fi
+        if [[ "$version" == "clean" || "$version" == "all" ]]; then
+            wget -O temp_16k_clean.tar.gz https://archive.org/download/hrsp-2mix-16000-real-human-clean/HRSP2mix_16000_real_human_clean.zip
+            tar -xzf temp_16k_clean.tar.gz -C HRSP2mix_16000_real_human/
+            rm temp_16k_clean.tar.gz
+        fi
     fi
 fi
